flask/app_engine/book_collection/requirements.txt
Flask==0.11.1
google-api-python-client==1.6.2

